+++ components/more-stories.tsx
@@ -1,5 +1,5 @@
-import Post from '../types/post';
 import PostPreview from './post-preview';
+import { Post } from '~/domains/Post';
 
 type Props = {
   posts: Post[];
