+++ components/post-preview.tsx
@@ -1,8 +1,8 @@
 import Link from 'next/link';
-import Author from '../types/author';
 import Avatar from './avatar';
 import DateFormatter from './date-formatter';
 import CoverImage from './cover-image';
+import Author from '~/domains/Author';
 
 type Props = {
   title: string;
