+++ src/pages/_app.tsx
@@ -1,6 +1,5 @@
 import { NextUIProvider } from '@nextui-org/react';
 import { AppProps } from 'next/app';
-import '../styles/index.css';
 
 export default function MyApp({ Component, pageProps }: AppProps) {
   return (
