+++ src/pages/index.tsx
@@ -1,8 +1,8 @@
 import Head from 'next/head';
 
 import Button from '@nextui-org/react/button';
-import { getAllPosts } from '../lib/api';
-import Post from '../types/post';
+import { getAllPosts } from '../../lib/api';
+import { Post } from '~/domains/Post';
 
 type Props = {
   allPosts: Post[];
