+++ src/pages/posts/[slug].tsx
@@ -1,24 +1,24 @@
 import { useRouter } from 'next/router';
 import ErrorPage from 'next/error';
 import Head from 'next/head';
-import Container from '../../components/container';
-import PostBody from '../../components/post-body';
-import Header from '../../components/header';
-import PostHeader from '../../components/post-header';
-import Layout from '../../components/layout';
-import { getPostBySlug, getAllPosts } from '../../lib/api';
-import PostTitle from '../../components/post-title';
-import { CMS_NAME } from '../../lib/constants';
-import markdownToHtml from '../../lib/markdownToHtml';
-import PostType from '../../types/post';
+import Container from '../../../components/container';
+import PostBody from '../../../components/post-body';
+import Header from '../../../components/header';
+import PostHeader from '../../../components/post-header';
+import Layout from '../../../components/layout';
+import { getPostBySlug, getAllPosts } from '../../../lib/api';
+import PostTitle from '../../../components/post-title';
+import { CMS_NAME } from '../../../lib/constants';
+import markdownToHtml from '../../../lib/markdownToHtml';
+import { Post } from '~/domains/Post';
 
 type Props = {
-  post: PostType;
-  morePosts: PostType[];
+  post: Post;
+  morePosts: Post[];
   preview?: boolean;
 };
 
-const Post = ({ post, preview }: Props) => {
+const PostPage = ({ post, preview }: Props) => {
   const router = useRouter();
   if (!router.isFallback && !post?.slug) {
     return <ErrorPage statusCode={404} />;
@@ -48,7 +48,7 @@ const Post = ({ post, preview }: Props) => {
   );
 };
 
-export default Post;
+export default PostPage;
 
 type Params = {
   params: {
