--- types/post.ts
@@ -1,16 +1,0 @@
-import Author from './author';
-
-type PostType = {
-  slug: string;
-  title: string;
-  date: string;
-  coverImage: string;
-  author: Author;
-  excerpt: string;
-  ogImage: {
-    url: string;
-  };
-  content: string;
-};
-
-export default PostType;
